Hopefully fixed the seg fault. Probably generating a png image too.

+++ vision/src/yuv422ToRGB.hpp
@@ -13,7 +13,7 @@ Every U and V bytes goes with 2 Y bytes. There is 2 * much brightness info than 
 
 #include <stdint.h>
 
-void yuv422_to_rgb32(const uint8_t *image_yuv422, uint8_t *rgbaData, int width, int height);
+void yuv422_to_rgb32(const uint8_t *image_yuv422, uint32_t *rgbaData, int width, int height);
 
 void yuv422_to_rgb32_int(uint8_t *image_yuv422, uint32_t *image_rgb32, int width, int height);
